Implement support for note off as note on w/ vel = 0
Implement support for sources which generate note on with velocity
0 as note off. Since in this case, all three bytes of the note on/off
message must be received before transmitting the message (since we
need to know if it is a note on or off in order for the transpose
memory to work), this is termed normal latency mode (960 us delay).

Switching between normal and low latency mode is automatic. When a note
on with velocity 0 is received, switches to normal latency mode.
When a note off (status 128) is received, switches to low latency mode.