Add mode to maximize running status
In order to keep latency at a minimum, add option to utilize running
status as much as possible, regardless of how much running status the
source uses. This is to minimize the potential latency.

With this turned off, the device attempts to mimic the running status
behavior of the source.