Sustain pedal emulation mode (low_latency_mode only)
In order to handle synths that don't manage the sustain pedal
properly (e.g. Waldorf Blofeld, which seems to miss sustain
pedal events arbitarily, implement a mode whereby sustain pedal
CC's are not sent on, but simply cause not offs not to be sent on
and flag them as sustained in the note descriptor list.

When a sustain pedal off message is received, note offs for all
note descriptors with the sustain bit set are sent.

This has a few implications.

First, if the transpose or source channel is changed when a note
is received, and that note already exists as playing in the
note descriptor list, send a note off for that note, as otherwise,
when the new note data overwrites the slot in the note descriptor
list, the original transpose and channel number will be overwritten,
and thus it's not possible send a note off at a later time. This
is because the note descriptor list refers to the notes played on the
source keyboard, hence there cannot be more than one entry per
key.

The note off cannot be sent immediately, as we are in the middle
of a note on message, potentially on a different channel than the
note off needs to be sent on. So queue the note off until the
velocity byte has been sent.

If RELEASE_ALL_NOTES_WHEN_CHANNEL_OR_OCTAVE_CHANGED is set, then
whenever a new note is received on a different channel or octave
selection than the previous note, all existing notes are released.
This is rather severe, and was originally implemented in order
to be easier to understand, but in fact results in the surprising
behavior that under the conditions described, all sustained notes
will be released when a new note is received. Thus this mode is
not recommented, and has not been fully tested.

Secondly, if the same note is played multiple times, this results
in several note ons being sent without intermediate note offs.
It seems that, with the synths tested so far (Blofeld and Zynthian,
with Obxd, Dexed, ZynAddSubFX, Pianoteq, NoizeMak3r, SF2) that a
single note off will in fact terminate all previous note ons receved
by the synth for the same key (the MiMi-a doesn't behave like this
I don't think though), so there's no risk of hanging notes with
this approach. However, it seems that it is customary to
assign a new voice for each note on received, which ultimately
causes the synth to run out of voices (on the Obxd this is especially
prominent as it doesn't seem to implement a voice robbing algorithm).
(It can result in a lovely swell effect if there are enough voices
though).

This second case has been be solved by, whenever a note is received
for the same note number and on the same channel and transpose as
a previous note, a note off is sent, _before_ sending the ongoing
note on (it cannot be sent after, as is the case with the
changed transpose/channel, as it would simply cut off the note on
just being sent!). Since the channel is the same, we can rely on the
same status message as is already being processed.

This is governed by the SEND_NOTE_OFF_FOR_EACH_SUSTAINED_NOTE