Add mode to stop sustaining when note channel changed
The usecase is something is being played on one channel,
with the sustain pedal held, then the source channel is
changed. With DONT_SUSTAIN_WHEN_NOTE_CHANNEL_CHANGED set,
the sustain status will be cleared, although the pedal is
not considered to have transitioned to the up state,
meaning that the notes on the previous channel will still
be sustained (logical, as the pedal is still down), but
not notes on the new channel (since the pedal has not been
pressed for this channel yet). This makes it possible to hold a
drone note, switch to another channel, still droning, without
the pedal considered to be held.

This should eventually be a switchable mode.