Add MIDI Dump feature, also refactoring display management
This feature dumps incoming or outgoing MIDI data, and displays
the 12 last received bytes in the right hand screen area,
using font size 1, in two columns. This is small font, so
hard to read, but does leave the rest of the display intact.

A parameter controls whether it is the MIDI input or output that
is dumped.

The data is not scrolled, but rather displayed in a wrapped-around
fashion, with a little dot indicating the last received byte.

Redrawing the screen for each received byte means that it delays
the processing of each MIDI byte, making the response sluggish,
and easily dropping data when too much arrives at once. Hence,
a new subclass Fast_SSD1306 for the display has been devised,
which not only prints to the screen memory but also directly to
the screen, alleviating the need for a separate display() call,
and at the same time in future making it possible to evolve
a display class which does not need the screen memory at all.

Limitations of current direct-to-screen implementation:
- Only I2C communication supported.
- Only characters can be printed, no lines, single pixels,
  fills, etc.
- Only white-on-black text.
- Only font sizes 1, 2, 3, 4 and 8 supported.
- Only prints character strings, no direct conversion
  from numbers. In order to minimize communication time
  with the display, it's advisable to put everything to be
  printed into a single buffer, which is then send to
  directPrint() for printing.
- Clear screen is implemented by writing spaces (3 of them)
  to the display with font size 8, which covers the whole
  display.
- The starting coordinates for the printout will have the y
  coordinate rounded down to the nearest multiple of 8.
- The font file (glcdfont.c) from Adafruit_GFX_Library is
  not available outside library, hence the file has been
  #included, meaning that currently there are two copies
  of the font data, wasting program memory space.

On the positive side, a noteworthy feature is that if a
string is printed which ends up outside the screen, any area
outside the screen is correctly disregarded without unexpected
wraparounds etc.

There seems to be considerable overhead when printing to the
display, much more than the bit rate of 1 MHz would imply.
For one thing, it seems that each so-called transaction
takes time in itself, so therefore the number of transactions
has been minimized by sending all the command bytes at the
start of the directPrint routine from a single buffer.

Even so, printouts can disrupt the MIDI data flow, so only print
out the latest byte(s) received, when the data actually arrives.

An optional mode defers printing until a set time (currently
50 ms) has passed with no incoming MIDI activity. However, this
could fail if the input data includes MIDI clock for instance
(however, that would cause the display to in practice fill up
with clock bytes, so MIDI dump would not be very useful in this
case anyway), so has been disabled.