Fix running status reset bug
The reset-running-status-every-second timeout sometimes
causes erroneous data, exacerbated when there are large delays
in the processing of MIDI data. Tests with a minimal echo
application reveal that there is no issue with the serial
interrupt.

With MIDI dump display enabled, it seems to happen fairly
often, but even without the dump display it will eventually
happen. The usual symptoms are a bunch of stuck notes that
were never played, i.e. it sounds like velocity bytes are
being mistaken for notes, or similar.

Turns out the problem is the potential double status byte
transmitted when in the NOTE_ON_NOTE state in low latency mode.
A status byte is emitted which if identical to the previous
status byte just transmitted would be handled by the
running status class; however, if running status is disabled
due to RUNNING_STATUS_TIMEOUT_US having been reached,
two identical status bytes will be output directly after each
other, causing certain devices, such as the Zynthian, to
misinterpret the incoming MIDI stream (in the case of the Zynthian,
note ons seem to be interpreted as note off and vice versa).