Implement Sustain Pedal No Running Status mode
The Waldorf Blofeld has a bug whereby sustain pedal messages
are lost arbitrarily, and upon closer investigation it turns
out that the synth does not respond properly to running status
for the sustain pedal. Thus, the sequence CC 64 0 64 127 is not
registered as the sustain pedal being down, and subsequent notes
are then not sustained. Pressing a key, then the pedal, works, as
then the pedal will trigger a new status message.

Fix this by always outputting a CC status message for the sustain
pedal, when the CC_PED_NRS mode is enabled.