Set up for Simple Box device
The Simple Box is a minimum-configuration MIDI octaver with added
modwheel support: it has no display, offering no other functionality
than octave shifting and generating modwheel data. Consequently
there are no parameters to set, hence the octaves go down to -3
instead of -2 with a settings mode at the leftmost octave switch
position. The potentiometer representing the mod wheel is connected
to pin A0.

The mode_flags are set to 0; in particular this means that the
no-pedal-running-status flag is not set.