Enable SKIP_CC mode, controlled by mode bits (default disabled)
Turns out that even when turning down the accelerometer
sensitivity on the Radikal Technologies Accelerator to min,
it still outputs CC 22 sometimes. This causes a Blofeld responding
to CC to set its LFO2 delay to 64 ... Since the disable
accelerometer function in the Accelerometer doesn't work,
we add that functionality here instead.

This function is disabled by default.