Set default (5.0V) analog reference at startup
When running on an Arduino Nano board using ATMEGA168 rather than
ATMEGA328, initially the AREF output was 0V. After setting the
analog deference to DEFAULT (= 5.0V AVCC), the AREF output was
5V as expected. Oddly though, this persisted even if the line
of code was subsequently removed, even after power cycling.
Anyway, it does no harm, so leave the line in.