Implement tracker interject timeout
If we receive a partial MIDI message (such as a note on with no
velocity byte), the tracker would hang, not allowing any modwheel
data to be interjected. To handle this, whenever the tracker state
does not allow interjection, give up after a second and reset it.

Only relevant when we want to interject (modwheel) data, hence
the #ifdef MODWHEEL.

In this condition, we set process_midi() to a new STATE_SKIP state,
which means that it will stop echoing data bytes until a new status
byte has arrived. This avoids spurious data bytes arriving after
the interject timeout has fired to be sent on, which would be
potentially considered to be data bytes utilizing running status
from the modwheel CC messasge.

The interject timeout is set from the last non-interjectable byte
sent rather than at the start of each interjectable state.

This mirrors what we want to do: When the state is non-interjectable,
we set a timeout from last byte sent in this state. Thus we
can cope with long sysex dumps, as long as the distance between
each individual byte sent is not overly long. Thus the interject
timeout can be rather short; currently set to 100 ms.