Avoid outputting CC value at startup
This can be due to two things: First, programmatically the
initial reference value needs to be the same as the value
read from the ADC, not 0 or other constant value, or else the
comparison with the 'previous value' will fail the first time,
resulting in the current value being transmitted. Secondly,
it seems a dummy ADC read is needed after starting it (possibly
because of the change of prescaler setting) to get rid of any
initial state.